# Neu_26-10-2026/openapi/forecast.yaml
openapi: 3.0.3
info:
  title: Phoenyra Forecast API
  version: 0.1.0
servers:
  - url: http://localhost:9500
paths:
  /forecast/dayahead:
    post:
      summary: Create 24h dayahead forecast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                area: { type: string, example: "AT" }
                assets:
                  type: array
                  items: { type: string }
                horizon_h: { type: integer, example: 24 }
                granularity_min: { type: integer, example: 15 }
                features:
                  type: object
                  properties:
                    weather: { type: boolean }
                    entsoe_price: { type: boolean }
                    entsoe_load: { type: boolean }
      responses:
        "202":
          description: Forecast job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
  /forecast/intraday:
    post:
      summary: Rolling intraday forecast 6-12h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                area: { type: string }
                horizon_h: { type: integer, example: 12 }
                granularity_min: { type: integer, example: 15 }
      responses:
        "202":
          description: Forecast job accepted
  /forecast/status/{job_id}:
    get:
      summary: Get forecast job status
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Status + results (if ready)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "running|done|failed" }
                  series:
                    type: array
                    items:
                      type: object
                      properties:
                        ts: { type: string, format: date-time }
                        price_eur_mwh: { type: number }
                        load_mw: { type: number }
                        gen_solar_mw: { type: number }
                        gen_wind_mw: { type: number }
                        soc_target_pct: { type: number }